Item Layout
System I began to use to place items on the screen for application use.  It uses a base frame size based on common smartphone size for initial placement, and augments the placement based on the size of the screen of the new device being used.  The number of values used is specific to the page being viewed.